Extract Brand to a separate file and document
