+++ packages/server-api/src/index.ts
@@ -9,6 +9,7 @@ export * from './course'
 export * from './autopilotapi'
 export * from './autopilotapi.guard'
 export * from './propertyvalues'
+export * from './brand'
 
 export interface Position {
   latitude: number
@@ -31,5 +32,3 @@ export enum SKVersion {
   v1 = 'v1',
   v2 = 'v2'
 }
-
-export type Brand<K, T> = K & { __brand: T }
